+++ src/main/java/studentmanagement/10001.txt
@@ -1,0 +1,5 @@
+courses enrolled 
+ --------------------------------------------------COSC110 - Introduction to Networking
+COSC102 - Computer Logics
+COSC101 - Introduction to Computer Science
+a fail not here:
