+++ src/main/java/studentmanagement/Lecturer.java
@@ -1,4 +1,4 @@
-package main.java.studentmanagement;
+package studentmanagement;
 public class Lecturer extends Person{
     private String name;
     public Lecturer(){}
