+++ src/main/java/studentmanagement/Student.java
@@ -1,4 +1,4 @@
-package main.java.studentmanagement;
+package studentmanagement;
 
 //imports
 import java.io.BufferedWriter;
