+++ src/main/java/studentmanagement/StudentGui.java
@@ -1,4 +1,5 @@
-package main.java.studentmanagement;
+package studentmanagement;
+
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
 import javax.swing.JFrame;
@@ -262,32 +263,7 @@ public class StudentGui extends JPanel {
     
 }
 
-class Reusable extends JFrame{
-
-    public  Reusable(String Message){
-        JFrame frame=new JFrame();
-        frame.setLayout(new BorderLayout());
-        frame.setSize(800,400);
-        frame.setDefaultCloseOperation(DISPOSE_ON_CLOSE);
-        frame.setVisible(true);
-
-        JTextField text=new JTextField();
-        text.setEditable(false);
-        frame.add(text,BorderLayout.CENTER);
-        text.setText(Message);
-
-        JButton button=new JButton("Close");
-        button.addActionListener(e->{
-            frame.dispose();
-            
-        });
-        frame.add(button,BorderLayout.SOUTH);
-
-    }
-}
-
-
- class CourseSelection {
+class CourseSelection {
     private JCheckBox[] checkBoxes;
     private JTextArea outputArea = new JTextArea(5, 30);
     private StringBuilder selectedCourses = new StringBuilder();
